requirements.txt
aiofiles
apscheduler
asyncio
google-api-python-client
google_auth_oauthlib
motor
oauth2client
requests
pyrogram
pymongo
python-dotenv
telegraph
tgcrypto
